[KV cache manager] No functional change intended, separate updateToken into addToken and removeToken
Given that streamLLM is broken now, removed related code logic and added
assertion to guard for sinkAttention settings. This allows us to revisit
the computation in the future.

The beam search is also broken now, removed related code logic and add
assertion and added assertion to guard beam search. Dead function
`replaceSharedBlock` is removed. We will add it back when supporting
beam search in the future.

Reasons of this change:
addToken is frequently checked, the two function should be decoupled.
Flattens callstack and nested-if.
std::swap use is bizarre.

Signed-off-by: eopXD <yuehtingc@nvidia.com>