[kv cache manager] Support SWA KV cache reuse
This merge request hopes to support SWA KV cache reuse. Before this
commit, when dealing with sliding window attention (SWA), the KV cache
manager allocates "window size" number of blocks and reuse them in a
cyclic matter. This behavior will not be able to support reuse since
block contents will be overwritten.

In this commit, we change such behavior to let the manager write blocks
in a linear matter. With a linear block writing bahavior, as the
attention window moves on, the out-of-window (OOW) blocks will be
detached. Right now for the sake of a correct feature first, we
directly offload the OOW block from the primary block pool (GPU memory)
to the secondary block pool (host memory). We will improve this in the
future by delegating the block movement to the eviction policy.

Signed-off-by: eopXD <yuehtingc@nvidia.com>
Co-authored-by: Tomer Asida <tasida@nvidia.com>